修复类型转换问题并设置最低平台版本
在 `GoofishDlg.cpp` 中显式添加按钮位置和间距的 `int` 类型转换，避免隐式类型转换问题。
在 `targetver.h` 中定义 `_WIN32_WINNT` 宏为 `0x0601`，明确最低支持的 Windows 平台版本为 Windows 7。

+++ Goofish/GoofishDlg.cpp
@@ -94,9 +94,9 @@ BOOL CGoofishDlg::OnInitDialog()
 		// 按钮尺寸和位置
 		const int btnWidth = m_uiModel->GetButtonWidth();
 		const int btnHeight = m_uiModel->GetButtonHeight();
-		const int btnTop = 10 * Utils::GetDpi();
-		const int btnSpacing = 10 * Utils::GetDpi();
-		const int btnLeftStart = 20 * Utils::GetDpi();
+		const int btnTop = (int)(10 * Utils::GetDpi());
+		const int btnSpacing = (int)(10 * Utils::GetDpi());
+		const int btnLeftStart = (int)(20 * Utils::GetDpi());
 
 		// 创建 Stop 按钮
 		m_btnStop.Create(_T("Stop"), WS_CHILD | WS_VISIBLE | BS_PUSHBUTTON,
